Refactor selector parse loop (#12)
Instead of using one pattern and using finditer, we use a dictionary
of pattern objects. Each pattern can report if it is available for a
specific document flags. This allows us to sequence the patterns and
avoid the patterns we want depending on the mode. This could allow for
easy custom stuff in the future.