Improve exception messages when a token is not found (#73)
In most cases, a full selector is parsed as a token, and if not, we
report the character we are at as invalid. While better context is
provided in exceptions for tokens that are fully parsed and validated,
when we don't parse a token, very little is conveyed back to the user.
There are only a few types of selectors allowed in CSS: tags, attr,
class, id, pseudo-class, pseudo-elements, and combinators. You can have
variations of these, but that is it.  We cover most of these cases
already with decent context, but there are still a few case that are
just reported as invalid characters. Moving forward, we will see if the
invalid character represents the start of one of the basic, supported
types. Only attr, class, id, and pseudo-(class|element) really fall into
the category that would generate one of these ambiguous exceptions, so
identify their starting character and raise an appropriate "malformed"
selector exception.

Also remove unnecessary "+" in PAT_PSEUDO_CLASS.