quirks mode (#84)
This aims to provide a transitional mode known as "quirks mode" to accommodate a couple non standard behaviors that Beautiful Soup used to allow. It breaks the CSS spec, but would allow Beautiful Soup to work on old scripts that exploited this behavior. This mode will not be documented, and when these quirks are detected, warnings will be raised to encourage users to fix their scripts. The idea is users should follow the CSS spec, which Soup Sieve will strictly enforce by default. Quirks mode would only be used by Beautiful Soup, and ideally will be weened off such dependency in the future. For this reason, as the documented API is not changing, we will consider this a bug fix release, unless it gets included with more major features.