Refactor of match (#86)
Refactor by splitting element related manipulation to its own mixin
class. Also move input validation and parsing to its own mixin. Ensure
we are calling the helper methods in the mixin.  This will help a few
areas where we are reusing the same logic over and over and help us be
more consistent. This may allow in the future the ability to derive
some of the classes to perform CSS selects on something like etree (or
other) objects directly if we feel the need in the future.