Skip tests that require html5lib or lxml when not available (#100)
* Skip tests that require html5lib or lxml

html5lib and lxml are optional bs4 dependencies. If they are not
available, test without them, skipping any tests that weren't able to be
run.

In Debian, we package separate cPython 2.x and PyPy 2.x stacks. bs4 is
available for PyPy in Debian, but html5lib and lxml aren't, yet. Let's
just skip these tests when we can't run them, as bs4 does for its own
tests.

* Add travis tests without optional bs4 dependencies

So that we don't regress support for them in the test-suite.

* Explicit flake8 ignores