Adjust traceback formatting of syntax errors
Now tracebacks that show a syntax error in a multiline selector will
look right:

      File "<string>", line 2
        input
           [type=3]
           ^
    soupsieve.css_parser.SelectorSyntaxError: Malformed attribute selector at position 9

There's perhaps a slight confusion when the error is not on the last
line, because the caret will be shown after the last line, like this:

      File "<string>", line 2
        input
           [type=3]
           [name=foo]
           ^
    soupsieve.css_parser.SelectorSyntaxError: Malformed attribute selector at position 9

I think it's acceptable: if the user pays attention to "line 2", they
can figure it out.

It might be better to change the indentation for the line with the error
to be a little arrow, like this:

      File "<string>", line 2
        input
    -->    [type=3]
           [name=foo]
           ^
    soupsieve.css_parser.SelectorSyntaxError: Malformed attribute selector at position 9

but I thought that might confuse Python traceback parsers, so I didn't
do it.