Backout REVERSE flag for now (#139)
Backing out REVERSE flag for now as it is uncertain what future changes
in relation to iframe we will make, and REVERSE may turn out to be
expensive if we have to account for such things. Rather not add it, and
see down the road that it won't be too bad, then add it now, and find
supporting it would be too painful.