HTML specific selector improvements
Ensure that all HTML specific selectors are comparing all related
elements against the XHTML namespace. In HTML parsers that do not
provide an HTML namespace for HTML elements, assume XHTML namespace.

Ensure that :lang(), :dir(), :default, and :indeterminate all consider
iframe boundaries and do not evaluate parent or child documents when
determining whether an element matches or not.

When a scoped selector is part of an iframe, :root should match that
scoped element's document root, and not the root of the whle tree.

Reference: #138