Shortcut to last descendant if available (#143)
If an element has that has content we wish to skip, we have to calculate
the last descendant. Usually this is done by descending down the various
`content` members of the tags looking at their last node until we find
one that is not a tag. But we can shortcut this if the element whose
content we wish to skip has a sibling.