Revert :disabled change (#157)
Browsers never let optgroup get nested as they shuffle elements around
and even strip out tags that don't belong under an optgroup if it is
under a select. For this reason, using the browser as an example is not
always the best unless the tree is also examined and rendering.

Ultimately, we don't need to go out of our way account for bad HTML, and
for the fact that the Python parsers don't quite parse exactly like a
browser, so we will continue to simply match according to the rules.