Initial work to canonicalize language tags and ranges
This work may not be used depending on what is decided in the level 4
selector spec, but if it is potentially done, we should hopefully be
ready. Tests are still needed to ensure we are doing things as expected.
Also, I am not entirely certain how ranges are supposed to be handled vs
tags, but I think I am handling them as expected.