Add experimental regex selector in :contains() and [attribute]
Regex selectors in the form /pattern/ can be used in :contains() and in
the value portion of [attribute].

They will follow the same case rules as set forth by those pseudo
selectors and attribute selectors, but you can alter the case with
regex flags or with the built-in attribute flags.

This is currently just an experiment. It is possible in this experiment
that we may expand this to tags, namespaces, and attribute names.

We are not committed to releasing this at this time and is only meant
for exploring possibilities, but this could turn into a real feature.