Improve pseudo-class error message
When encountering a pseudo-class that is unrecognized, we should not
state it is "unsupported" a it may be entirely invalid. While there are
some pseudo-classes that are valid but not supported, the status of such
things can change all the time and our internal list may become
outdated.

Additionally, sometimes users trying to incorrectly specify tags
with namespaces can trigger this message as they may assume that because
a namespace is declared in their document as `ns:element` that they can
do this in CSS as well. This specific case is not easily detectable for
us as we generically parse the syntax, not the intent of the syntax.

To clear up confusion, be more specific and simply state that anytime
a pseudo-class is not recognized by soupsieve that it is either invalid
or not recognizable by soupsieve. By stating both possibilities, this
points out that pseudo-class syntax was noted and that it is
unrecognized by soupsieve, regardless of status, and may be entirely
invalid. Additionally, note that if it being recognized as a
pseudo-class is a surprise that the colon can be escaped to avoid
the recognition as such.